refactor(github-client): reorganize client into lifecycle folders
