+++ src/main/github/client/lookup/pr-branch-lookup.ts
@@ -1,0 +1,155 @@
+import { ghExecFileAsync, ghRepoExecOptions, type OwnerRepo } from '../../gh-utils'
+import { githubHostExecOptions, type GitHubApiRepository } from '../../github-api-repository'
+import type { GhExecOptions } from './../github-exec-scope'
+import { isNoPullRequestError } from './../gh-error-predicates'
+import { PR_LOOKUP_JSON_FIELDS, PR_BRANCH_LIST_JSON_FIELDS, mapRestPullRequest, normalizePullRequestLookupData, type PullRequestLookupData, type RestPullRequest } from './pull-request-lookup-data'
+import { getPRByNumber } from './pr-number-lookup'
+export async function getRestPRForBranch(
+  prRepo: GitHubApiRepository,
+  headOwner: string,
+  branchName: string,
+  ghOptions: ReturnType<typeof ghRepoExecOptions>
+): Promise<PullRequestLookupData | null> {
+  const head = encodeURIComponent(`${headOwner}:${branchName}`)
+  const { stdout } = await ghExecFileAsync(
+    ['api', `repos/${prRepo.owner}/${prRepo.repo}/pulls?head=${head}&state=all&per_page=1`],
+    { ...ghOptions, ...githubHostExecOptions(prRepo) }
+  )
+  const list = JSON.parse(stdout) as RestPullRequest[]
+  const pr = list[0]
+  return pr ? mapRestPullRequest(pr) : null
+}
+
+export async function getFallbackPRListForBranch(
+  prRepo: GitHubApiRepository,
+  branchName: string,
+  ghOptions: ReturnType<typeof ghRepoExecOptions>
+): Promise<PullRequestLookupData | null> {
+  const { stdout } = await ghExecFileAsync(
+    [
+      'pr',
+      'list',
+      '--repo',
+      `${prRepo.owner}/${prRepo.repo}`,
+      '--head',
+      branchName,
+      '--state',
+      'all',
+      '--limit',
+      '1',
+      '--json',
+      PR_BRANCH_LIST_JSON_FIELDS
+    ],
+    { ...ghOptions, ...githubHostExecOptions(prRepo) }
+  )
+  const list = JSON.parse(stdout) as PullRequestLookupData[]
+  return list[0] ?? null
+}
+
+export async function hydrateBranchLookupWithExactPR(
+  ownerRepo: OwnerRepo,
+  branchData: PullRequestLookupData | null,
+  ghOptions: GhExecOptions,
+  executionScope: string
+): Promise<PullRequestLookupData | null> {
+  if (!branchData) {
+    return null
+  }
+  try {
+    return (
+      (await getPRByNumber(ownerRepo, branchData.number, ghOptions, executionScope, branchData)) ??
+      branchData
+    )
+  } catch {
+    return branchData
+  }
+}
+
+export async function lookupPRByBranchName(args: {
+  candidates: OwnerRepo[]
+  headRepo: OwnerRepo | null
+  branchName: string
+  ghOptions: GhExecOptions
+  executionScope: string
+}): Promise<{
+  data: PullRequestLookupData | null
+  dataRepo: OwnerRepo | null
+  pendingError?: unknown
+}> {
+  if (args.candidates.length > 0) {
+    let pendingError: unknown
+    let hasPendingError = false
+    for (const candidate of args.candidates) {
+      try {
+        const branchData = args.headRepo
+          ? await getRestPRForBranch(
+              candidate,
+              args.headRepo.owner,
+              args.branchName,
+              args.ghOptions
+            )
+          : await getFallbackPRListForBranch(candidate, args.branchName, args.ghOptions)
+        // Why: REST/list branch lookup identifies the PR cheaply; exact `gh pr view` carries review, merge-queue, and auto-merge state.
+        const data = await hydrateBranchLookupWithExactPR(
+          candidate,
+          branchData,
+          args.ghOptions,
+          args.executionScope
+        )
+        if (data) {
+          return { data, dataRepo: candidate }
+        }
+      } catch (err) {
+        if (args.headRepo) {
+          throw err
+        }
+        if (!hasPendingError) {
+          pendingError = err
+          hasPendingError = true
+        }
+        try {
+          const branchData = await getRestPRForBranch(
+            candidate,
+            candidate.owner,
+            args.branchName,
+            args.ghOptions
+          )
+          const data = await hydrateBranchLookupWithExactPR(
+            candidate,
+            branchData,
+            args.ghOptions,
+            args.executionScope
+          )
+          if (data) {
+            return { data, dataRepo: candidate }
+          }
+        } catch (retryErr) {
+          if (!hasPendingError) {
+            pendingError = retryErr
+            hasPendingError = true
+          }
+        }
+      }
+    }
+    // Why: branch-list failures are ambiguous for fork discovery; give exact fallback-number recovery a chance before surfacing the error.
+    return hasPendingError
+      ? { data: null, dataRepo: null, pendingError }
+      : { data: null, dataRepo: null }
+  }
+
+  try {
+    const { stdout } = await ghExecFileAsync(
+      ['pr', 'view', args.branchName, '--json', PR_LOOKUP_JSON_FIELDS],
+      args.ghOptions
+    )
+    return {
+      data: normalizePullRequestLookupData(JSON.parse(stdout) as PullRequestLookupData),
+      dataRepo: null
+    }
+  } catch (err) {
+    if (isNoPullRequestError(err)) {
+      return { data: null, dataRepo: null }
+    }
+    throw err
+  }
+}
