+++ src/main/github/client/lookup/pull-request-lookup-hydration.ts
@@ -1,0 +1,30 @@
+import { type OwnerRepo } from '../../gh-utils'
+import type { GhExecOptions } from './../github-exec-scope'
+import { detectRepositoryMergeMetadata } from './../detect/repository-merge-metadata'
+import { normalizePullRequestLookupData, type PullRequestLookupData } from './pull-request-lookup-data'
+export async function hydratePullRequestLookupData(
+  ownerRepo: OwnerRepo,
+  data: PullRequestLookupData,
+  ghOptions: GhExecOptions,
+  executionScope: string
+): Promise<PullRequestLookupData> {
+  const normalized = normalizePullRequestLookupData(data)
+  const hasRichMergeFields =
+    'reviewDecision' in data || 'mergeStateStatus' in data || 'autoMergeRequest' in data
+  const mergeMetadata = hasRichMergeFields
+    ? await detectRepositoryMergeMetadata(
+        ownerRepo,
+        normalized.stack?.baseRefName ?? normalized.baseRefName,
+        ghOptions,
+        executionScope
+      )
+    : undefined
+  return {
+    ...normalized,
+    ...(mergeMetadata ? { mergeQueueRequired: mergeMetadata.mergeQueueRequired } : {}),
+    ...(mergeMetadata ? { autoMergeAllowed: mergeMetadata.autoMergeAllowed } : {}),
+    ...(mergeMetadata?.mergeMethodSettings
+      ? { mergeMethodSettings: mergeMetadata.mergeMethodSettings }
+      : {})
+  }
+}
